add `env` module for typing environment
sdiehl has a module which presents a subset of (the wrapped)
`Data.Map` API.

this is my take on that. it doesn't seems strictly necessary - we can
probably just use `HashMap` directly, and perhaps it would be nicer to
just do that, because it implements so many handy traits.

I thought this would teach me a bit about API design and also about
Rust trait stuff.

we can rewrite later if we feel it's not worth its maintenance cost.