disallow trailing characters on keywords; avoid unnecessary attempts
I was previously forcing the presence of ' ' after keywords. now we
ensure the absence of alpha_num after them. this seems tidier.

previously we could encounter weird behavior like `(if 1)` parsing to a
function application with variable named `if`. this is because we would
backtrack if `if` failed to parse, and end up in the `App` branch. now I
avoid backtracking by committing (no `attempt`). in order todo so I had
to factor out the common prefix of lam & let ("l"). not sure if there's
a better way.