document PrimOp trickery & fix bug
I was mistakenly not checking from initial "enclosure" of PrimOps inside
of `App`s. so I would recognize all PrimOps as being enclosed... whoops.

also, in the case of PrimOp "lifting" (packaging into a closure), no
environment is needed because no external environment is needed in the
closure, just variable bindings which are already managed.