add `nil`, revamp `list`
revamp `list` to be just a syntactic form, implemented in terms of
`cons` and `nil`. this probably reduce efficiency of interpretation,
because we had a vectorized `list` before, which packaged everything up
into a vector in one pass, with no intermediate closures.

I haven't heavily tested this but it seems to work...